% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedBlock.R
\name{FixedBlock}
\alias{FixedBlock}
\title{Permuted fixed block randomization}
\description{
A class for permuted fixed block randomization.
}
\section{Super class}{
\code{\link[randomization4trial:Randomization]{randomization4trial::Randomization}} -> \code{FixedBlock}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{block}}{vector of treatments to be included in each block}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedBlock-validate}{\code{FixedBlock$validate()}}
\item \href{#method-FixedBlock-new}{\code{FixedBlock$new()}}
\item \href{#method-FixedBlock-summary}{\code{FixedBlock$summary()}}
\item \href{#method-FixedBlock-clone}{\code{FixedBlock$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedBlock-validate"></a>}}
\if{latex}{\out{\hypertarget{method-FixedBlock-validate}{}}}
\subsection{Method \code{validate()}}{
Validate input parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedBlock$validate(block)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block}}{vector of treatments to be included in each block}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedBlock-new"></a>}}
\if{latex}{\out{\hypertarget{method-FixedBlock-new}{}}}
\subsection{Method \code{new()}}{
Create a new exponential distribution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedBlock$new(block)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block}}{vector of treatments to be included in each block}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Exponential} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedBlock-summary"></a>}}
\if{latex}{\out{\hypertarget{method-FixedBlock-summary}{}}}
\subsection{Method \code{summary()}}{
summarize the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedBlock$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedBlock-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FixedBlock-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedBlock$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
