% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomization.R
\name{Randomization}
\alias{Randomization}
\title{Randomization class}
\description{
Randomization class

Randomization class
}
\details{
A class for randomization methods in clinical trials
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Randomization-validate}{\code{Randomization$validate()}}
\item \href{#method-Randomization-assign}{\code{Randomization$assign()}}
\item \href{#method-Randomization-summary}{\code{Randomization$summary()}}
\item \href{#method-Randomization-clone}{\code{Randomization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Randomization-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Randomization-validate}{}}}
\subsection{Method \code{validate()}}{
validate input parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Randomization$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Randomization-assign"></a>}}
\if{latex}{\out{\hypertarget{method-Randomization-assign}{}}}
\subsection{Method \code{assign()}}{
generate assignments of subjects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Randomization$assign()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Randomization-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Randomization-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Randomization$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Randomization-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Randomization-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Randomization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
