% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization4trial-package.R
\docType{package}
\name{randomization4trial-package}
\alias{randomization4trial}
\alias{randomization4trial-package}
\title{randomization4trial: Randomization methods in clinical trials with R6}
\description{
The goal of 'randomization4trial' is to enable randomization methods in clincial trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/littlebeannie/randomization4trial}
  \item Report bugs at \url{https://github.com/littlebeannie/randomization4trial/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Other contributors:
\itemize{
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
