% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{Simulate stratified time-to-event outcome randomized trial}
\usage{
randomize(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Additional arguments (not used).}
}
\value{
TBA
}
\description{
Simulate stratified time-to-event outcome randomized trial
}
